+++ package.json
@@ -41,12 +41,6 @@
     "eject": "react-scripts eject",
     "server": "json-server --watch ./public/database/index.js --port 8000"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
   "browserslist": {
     "production": [
       ">0.2%",
@@ -69,21 +63,7 @@
     "@types/react-dom": "^18.0.0",
     "@types/store": "^2.0.2",
     "babel-eslint": "^10.1.0",
-    "eslint": "^8.14.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-config-react": "^1.1.7",
-    "eslint-config-react-app": "^7.0.1",
-    "eslint-plugin-flowtype": "^8.0.3",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.29.4",
-    "eslint-plugin-react-hooks": "^4.5.0",
-    "eslint-plugin-testing-library": "^5.3.1",
-    "eslint-webpack-plugin": "^3.1.1",
     "postcss-scss": "^4.0.4",
-    "prettier": "^2.7.1",
     "sass": "^1.53.0",
     "sass-loader": "^12.6.0",
     "style-loader": "^3.3.1",
